fixup! fixup! fixup! fixup! remove SMTP tests and infrastructure
